Changed button definition location.
The definition for the buttons now comes from a file dedicated to that
job - buttons.jsonc. Eventually I want this to be in a configurable
location, and for that file to be watched by the extension for changes.
This commit also contains a jsonValidation contribution to aid the
creation of the file.